SunGenerator class in Field and GeneratorPlants
Sun generation between Field and Plants were not consolidated. Player
storing the sun points will be a problem for deep copying Level later as
Level does not have access to Player.

Field now contains the sun data. The SunGenerator observes the Level
class and updates the Field automatically each turn. It has it's own
internal cooldown which triggers each time it generates sun and won't
generate sun again until the cooldown reaches 0.

Sun production on Field appears to be functional, however a bug has a
appeared where a unknown zombie appears to be trying to hit a plant but
is causing a NullPointerException since no plant exists on the square.